This fixes the crash in sentiment.Restore()
The crash in sentiment.Restore() is caused by the fact that
The tokenizer is not restored in goml after Unmarshal from json
This is likely not the best place to make this change as the issue
is in the goml library. A couple better fixes should likely be done.

1) Privide a Marshal/Unmarshal function as part of the NaiveBayes type
that would restore the tokenizer

2) Make the spaceTokenizer public. This would avoid the duplication of
code this change makes but would mean a manual process for restoring the
tokenizer after Unmarshaling

I perfer the former if possible. Unless it's intended to wrap all the json
functionality up somewhere elase and pull all the json references out of
bayes.go then bayes.go should be able to Marshal/Unmarshal with out
leading to a crash. This would mean also changing sentiment to accomidate that.

I went for this quick and dirty change to keep moving, open up discustion,
trigger the proper change, and avoid making dependant changes in multiple
repos.